End to this madness, better design of aggregate structs
